requirements.txt
appdirs==1.4.4
asgiref==3.7.2
audioread==3.0.0
backports.zoneinfo==0.2.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
decorator==5.1.1
Django==4.2.2
django-cors-headers==4.0.0
djangorestframework==3.14.0
idna==3.4
importlib-metadata==6.6.0
joblib==1.2.0
lazy-loader==0.2
librosa==0.10.0.post2
llvmlite==0.40.0
msgpack==1.0.5
numba==0.57.0
numpy==1.24.3
packaging==23.1
pooch==1.6.0
pycparser==2.21
pytz==2023.3
requests==2.31.0
scikit-learn==1.2.2
scipy==1.10.1
soundfile==0.12.1
soxr==0.3.5
sqlparse==0.4.4
threadpoolctl==3.1.0
typing-extensions==4.6.3
urllib3==2.0.2
zipp==3.15.0
